Fix Issue #204 "Can't create disks > 2GiB"
Change type of size field in definition and usage

+++ govcd/api_vcd_test.go
@@ -106,8 +106,8 @@ type TestConfig struct {
 		ExternalNetworkPortGroupType string `yaml:"externalNetworkPortGroupType,omitempty"`
 		VimServer                    string `yaml:"vimServer,omitempty"`
 		Disk                         struct {
-			Size          int `yaml:"size,omitempty"`
-			SizeForUpdate int `yaml:"sizeForUpdate,omitempty"`
+			Size          int64 `yaml:"size,omitempty"`
+			SizeForUpdate int64 `yaml:"sizeForUpdate,omitempty"`
 		}
 	} `yaml:"vcd"`
 	Logging struct {
